updateBattlefield Methode nach Battlefield Klasse verschoben
