+++ src/main/java/com/btcag/bootcamp/Main.java
@@ -2,7 +2,7 @@ package com.btcag.bootcamp;
 
 import java.util.Scanner;
 
-import static com.btcag.bootcamp.Battlefield.createBattlefield;
+import static com.btcag.bootcamp.Battlefield.*;
 import static com.btcag.bootcamp.Game.*;
 import static com.btcag.bootcamp.Mechanics.enemyMove;
 import static com.btcag.bootcamp.Mechanics.playerMove;
@@ -79,23 +79,6 @@ public class Main {
             } else {
                 System.out.println("Diese Eingabe ist ungültig, geben Sie bitte neu ein!");
             }
-        }
-    }
-
-    //Updates Battlefield on users interface
-    public static void updateBattlefield(Battlefield battlefield, Character player, Character opponent) {
-        System.out.println();
-        for (int i = 0; i < battlefield.hight; i++) {
-            for (int j = 0; j < battlefield.width; j++) {
-                if (i == player.getPositionX() && j == player.getPositionY()) {
-                    System.out.print(player.getChar());
-                } else if (i == opponent.getPositionX() && j == opponent.getPositionY()) {
-                    System.out.print(opponent.getChar());
-                } else {
-                    System.out.print(battlefield.battlefieldArray[i][j] + " ");
-                }
-            }
-            System.out.println();
         }
     }
 }
